+++ tests/Fixture/Templates/TestController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Skeleton\Http\Controllers;
+
+/**
+ * TestController
+ *
+ * @package Skeleton
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class TestController
+{
+    //
+}